Todyl Integration with VPC Flow Logs
This script is setup to be run AFTER you run the CF script to ingest AWS CloudTrail data.

Run this integration script in CloudFormation for each VPC you want to monitor in Todyl.  Check the Outputs from the Stack job to get the Queue URL.  Then go to the Outputs from the CloudTrail integration stack in CF and get the Access Key ID and Secret Access Key values and paste them into the Todyl config.